requirements.txt
discord.py
python-dotenv
supabase
pytz
requests
openai
coverage
codecov

kluvs-brain @ git+https://github.com/ivangarzab/kluvs-brain.git